Create usmap/api.js and events.js
Create ajax method for item GET action. Move the usMap function from
index.js to newly created usmap/events.js file. This sets us up for a
modular file structure moving forward -- next step is to hook up ajax
request to click event and filter responses according to which state was
clicked.
